refactor: split plugin/tests.rs into focused test modules
Break 607-line monolithic plugin test file into five focused submodules:
- manifest: TOML parsing tests (4 tests)
- validation: manifest validation rule tests (6 tests)
- discovery: filesystem plugin discovery tests (4 tests)
- runtime: plugin init/shutdown lifecycle tests (6 tests)
- status: persistence-backed status snapshot tests (4 tests)

All 24 existing tests preserved with identical assertions.
